+++ src/application/t2d/modules/flattened_dataset_description.py
@@ -1,0 +1,49 @@
+from pathlib import Path
+
+from psycop_feature_generation.data_checks.flattened.data_integrity import (
+    save_feature_set_integrity_from_dir,
+)
+from psycop_feature_generation.data_checks.flattened.feature_describer import (
+    save_feature_description_from_dir,
+)
+
+
+def save_feature_set_description_to_disk(
+        predictor_specs: list,
+        flattened_dataset_file_dir: Path,
+        out_dir: Path,
+        file_suffix: str,
+        describe_splits: bool = True,
+        compare_splits: bool = True,
+):
+    """Describe output.
+
+    Args:
+        predictor_specs (list): List of predictor specs.
+        flattened_dataset_file_dir (Path): Path to dir containing flattened time series files.
+        out_dir (Path): Path to output dir.
+        file_suffix (str): File suffix.
+        describe_splits (bool, optional): Whether to describe each split. Defaults to True.
+        compare_splits (bool, optional): Whether to compare splits, e.g. do all categories exist in both train and val. Defaults to True.
+    """
+
+    # Create data integrity report
+    if describe_splits:
+        save_feature_description_from_dir(
+            feature_set_dir=flattened_dataset_file_dir,
+            feature_specs=predictor_specs,
+            splits=["train"],
+            out_dir=out_dir,
+            file_suffix=file_suffix,
+        )
+
+    # Describe/compare splits control flow happens within this function
+    if compare_splits:
+        save_feature_set_integrity_from_dir(
+            feature_set_dir=flattened_dataset_file_dir,
+            split_names=["train", "val", "test"],
+            out_dir=out_dir,
+            file_suffix=file_suffix,
+            describe_splits=describe_splits,
+            compare_splits=compare_splits,
+        )
