+++ src/application/t2d/modules/save_dataset_to_disk.py
@@ -1,0 +1,71 @@
+from pathlib import Path
+from typing import Optional
+
+import numpy as np
+import pandas as pd
+
+import psycop_feature_generation.loaders
+from psycop_feature_generation.utils import write_df_to_file
+
+
+def split_and_save_dataset_to_disk(
+        flattened_df: pd.DataFrame,
+        out_dir: Path,
+        file_prefix: str,
+        file_suffix: str,
+        split_ids_dict: Optional[dict[str, pd.Series]] = None,
+        splits: Optional[list[str]] = None,
+):
+    """Split and save to disk.
+
+    Args:
+        flattened_df (pd.DataFrame): Flattened dataframe.
+        out_dir (Path): Path to output directory.
+        file_prefix (str): File prefix.
+        file_suffix (str, optional): Format to save to. Takes any of ["parquet", "csv"].
+        split_ids_dict (Optional[dict[str, list[str]]]): Dictionary of split ids, like {"train": pd.Series with ids}.
+        splits (list, optional): Which splits to create. Defaults to ["train", "val", "test"].
+    """
+
+    if splits is None:
+        splits = ["train", "val", "test"]
+
+    msg = Printer(timestamp=True)
+
+    flattened_df_ids = flattened_df["dw_ek_borger"].unique()
+
+    # Version table with current date and time
+    # prefix with user name to avoid potential clashes
+
+    # Create splits
+    for dataset_name in splits:
+        if split_ids_dict is None:
+            df_split_ids = psycop_feature_generation.loaders.raw.load_ids(
+                split=dataset_name,
+            )
+        else:
+            df_split_ids = split_ids_dict[dataset_name]
+
+        # Find IDs which are in split_ids, but not in flattened_df
+        split_ids = df_split_ids["dw_ek_borger"].unique()
+        flattened_df_ids = flattened_df["dw_ek_borger"].unique()
+
+        ids_in_split_but_not_in_flattened_df = split_ids[
+            ~np.isin(split_ids, flattened_df_ids)
+        ]
+
+        msg.warn(
+            f"{dataset_name}: There are {len(ids_in_split_but_not_in_flattened_df)} ({round(len(ids_in_split_but_not_in_flattened_df) / len(split_ids) * 100, 2)}%) ids which are in {dataset_name}_ids but not in flattened_df_ids, will get dropped during merge. If examining patients based on physical visits, see 'OBS: Patients without physical visits' on the wiki for more info.",
+        )
+
+        split_df = pd.merge(flattened_df, df_split_ids, how="inner", validate="m:1")
+
+        # Version table with current date and time
+        filename = f"{file_prefix}_{dataset_name}.{file_suffix}"
+        msg.info(f"Saving {filename} to disk")
+
+        file_path = out_dir / filename
+
+        write_df_to_file(df=split_df, file_path=file_path)
+
+        msg.good(f"{dataset_name}: Succesfully saved to {file_path}")
