+++ src/application/t2d/overview.md
@@ -1,0 +1,10 @@
+1. Misc. setup
+   Init wandb
+
+2. Feature specification
+
+3. Flattening
+
+4. Saving dataset to disk
+
+5. Feature evaluation
