+++ src/psycop_feature_generation/loaders/example/load_admissions.py
