+++ src/psycop_feature_generation/loaders/example/load_bmi.py
