+++ src/psycop_feature_generation/loaders/example/load_structured_sfi.py
